Refactor fix
